Move test to test_free_threading.test_sys
