+++ Lib/test/test_sys.py
@@ -1369,20 +1369,6 @@ class SysModuleTest(unittest.TestCase):
         with self.assertRaises(TypeError):
             sys.set_int_max_str_digits(2_048.0)
 
-    def test_int_max_str_digits_thread(self):
-        # gh-151218: Check that it's safe to call set_int_max_str_digits()
-        # in parallel. Previously, this test triggered warnings in TSan
-        # on a free threaded build.
-
-        old_limit = sys.get_int_max_str_digits()
-        self.addCleanup(sys.set_int_max_str_digits, old_limit)
-
-        def worker():
-            for i in range (20_000):
-                sys.set_int_max_str_digits(4300 + (i & 7))
-
-        threading_helper.run_concurrently(worker, nthreads=4)
-
 
 @test.support.cpython_only
 @test.support.force_not_colorized_test_class
